refactor: fe tweaks to consume conerns api
TISNEW-4597
