+++ src/app/shared/records/constants.ts
@@ -1,4 +1,10 @@
+import { Sort } from "@angular/material/sort";
 import { IRecordDataCell } from "./records.interfaces";
+
+export const DEFAULT_SORT: Sort = {
+  active: "submissionDate",
+  direction: "desc"
+};
 
 export const RECORDS_COLUMN_DATA: IRecordDataCell[] = [
   {
