+++ src/app/trainees/constants.ts
@@ -1,10 +1,3 @@
-import { Sort } from "@angular/material/sort";
-
-export const DEFAULT_SORT: Sort = {
-  active: "submissionDate",
-  direction: "desc"
-};
-
 export const COLUMN_DATA: any[][] = [
   ["GMC Submission due date", "submissionDate", true],
   ["Status", "doctorStatus", false],
